LayoutTests/imported/w3c/web-platform-tests/html/syntax/speculative-parsing/generated/document-write/svg-script-src.tentative.sub-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: Error: assert_equals: speculative case incorrectly fetched expected "" but got "param-encodingcheck: %C4%9E\r\nAccept: */*\r\nReferer: http://web-platform.test:8800/html/syntax/speculative-parsing/gen...

    <!-- speculative case in document.write -->
    <svg><script src="/html/syntax/speculative-parsing/resources/stash.py?action=put&amp;uuid=c3305dfc-c350-43c4-8744-24cc3e1f4a37&amp;encodingcheck=&Gbreve;"></script></svg>


FAIL Speculative parsing, document.write(): svg-script-src Unhandled rejection: assert_equals: speculative case incorrectly fetched expected "" but got "param-encodingcheck: %C4%9E\r\nAccept: */*\r\nReferer: http://web-platform.test:8800/html/syntax/speculative-parsing/generated/document-write/svg-script-src.tentative.sub.html"
